requirements.txt
websocket-client==1.6.4
pandas==2.1.4
sqlalchemy==2.0.23
pyodbc==5.0.1
requests==2.31.0
python-dotenv==1.0.0
psutil==5.9.0
